CLN: include new module for BasePlot class, with optional color selection as standard
